backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
python-dotenv==1.0.1

# ML
torch>=2.3.0
torchvision>=0.18.0
facenet-pytorch==2.6.0
scikit-learn==1.5.1
numpy>=1.26.0
Pillow>=10.3.0
pillow-heif>=0.16.0
scipy>=1.13.0

# Google Drive
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.131.0

# Telegram
aiogram==3.10.0

# LLM
anthropic>=0.30.0
